test(app): split naval panel test files under 800-line gate (#4224)
Extract combine scenario tables and ScopedNavalPanelHarness so
repo.app_test_file_size passes after Slice D harness densification.
